refactor: test case
